+++ packages/Angular/Generic/tasks/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @memberjunction/ng-tasks
 
+## 3.0.0
+
+### Patch Changes
+
+- @memberjunction/ai-engine-base@3.0.0
+- @memberjunction/ai-core-plus@3.0.0
+- @memberjunction/core@3.0.0
+- @memberjunction/core-entities@3.0.0
+
 ## 2.133.0
 
 ### Patch Changes