+++ packages/Angular/Generic/timeline/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log - @memberjunction/ng-timeline
 
+## 3.0.0
+
+### Patch Changes
+
+- @memberjunction/core@3.0.0
+
 ## 2.133.0
 
 ### Patch Changes