+++ packages/Communication/engine/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log - @memberjunction/communication-engine
 
+## 3.0.0
+
+### Patch Changes
+
+- @memberjunction/communication-types@3.0.0
+- @memberjunction/core@3.0.0
+- @memberjunction/core-entities@3.0.0
+- @memberjunction/global@3.0.0
+- @memberjunction/templates@3.0.0
+
 ## 2.133.0
 
 ### Patch Changes