+++ packages/Communication/entity-comm-server/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log - @memberjunction/entity-communications-server
 
+## 3.0.0
+
+### Patch Changes
+
+- @memberjunction/communication-types@3.0.0
+- @memberjunction/communication-engine@3.0.0
+- @memberjunction/entity-communications-base@3.0.0
+- @memberjunction/core@3.0.0
+- @memberjunction/core-entities@3.0.0
+- @memberjunction/global@3.0.0
+
 ## 2.133.0
 
 ### Patch Changes