+++ packages/TestingFramework/CLI/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @memberjunction/testing-cli
 
+## 3.0.0
+
+### Patch Changes
+
+- @memberjunction/core@3.0.0
+- @memberjunction/core-entities@3.0.0
+- @memberjunction/global@3.0.0
+- @memberjunction/sqlserver-dataprovider@3.0.0
+- @memberjunction/testing-engine@3.0.0
+
 ## 2.133.0
 
 ### Patch Changes